rename and document RecordingMiddleware
